feat: setup stripe subscriptions
The abstractions are leaky though. Unless you add the concepts of a customer/id and price_id to the basic subscription domain.

app/subscription is for most subscription apis, but there's also app/subscription_portal which just sends users to stripe to handle checkout.

Webhooks are the same either way, but if developing the UI in-house, you wouldn't use app/subscription_portal services.